+++ tests/phpunit/Integration/Api/IncrementTest.php
@@ -27,10 +27,6 @@ use ApiTestCase;
 class IncrementTest extends ApiTestCase {
 
 	public function testIncrement() {
-		$this->markTestSkipped(
-			'Need to fix the "no such table: unittest_idprovider_increments" which occurs in test setting only'
-		);
-
 		$getId = function ( $prefix ) {
 			return $this->doApiRequest( [
 				'action' => 'idprovider-increment',